chore: move `segments.cpp` to `src/level/segments.cpp`
